+++ tests/include/tests/pbkdf2/pbkdf2.h
@@ -3,6 +3,6 @@
 
 #include <testenv.h>
 
-int test_pbkdf2(char *output, int maxlen, FILE *ext);
+int test_pbkdf2(char *output, size_t maxlen, FILE *ext);
 
 #endif
