+++ tests/include/tests/stream/stream.h
@@ -3,7 +3,7 @@
 
 #include <testenv.h>
 
-int test_stream(char *output, int maxlen, FILE *ext);
-int test_stream_utilities(char *output, int maxlen, FILE *ext);
+int test_stream(char *output, size_t maxlen, FILE *ext);
+int test_stream_utilities(char *output, size_t maxlen, FILE *ext);
 
 #endif
