+++ tests/include/tests/utility/errors.h
@@ -3,6 +3,6 @@
 
 #include <testenv.h>
 
-int test_error_codes(char *output, int maxlen, FILE *ext);
+int test_error_codes(char *output, size_t maxlen, FILE *ext);
 
 #endif
