+++ tests/include/tests/utility/utils.h
@@ -3,12 +3,12 @@
 
 #include <testenv.h>
 
-int test_macros(char *output, int maxlen, FILE *ext);
+int test_macros(char *output, size_t maxlen, FILE *ext);
 
-int test_pad_check(char *output, int maxlen, FILE *ext);
+int test_pad_check(char *output, size_t maxlen, FILE *ext);
 
-int test_xor_buffer(char *output, int maxlen, FILE *ext);
+int test_xor_buffer(char *output, size_t maxlen, FILE *ext);
 
-int test_inc_buffer(char *output, int maxlen, FILE *ext);
+int test_inc_buffer(char *output, size_t maxlen, FILE *ext);
 
 #endif
