+++ tests/src/tests/misc/os_random.c
@@ -2,7 +2,7 @@
 
 #include <misc/os_random.h>
 
-int test_os_random(char *output, int maxlen, FILE *ext)
+int test_os_random(char *output, size_t maxlen, FILE *ext)
 {
     uint8_t buffer[1024];
     int t;
